+++ scripts/4.2_create_new.py
@@ -6,19 +6,17 @@ from compas.geometry import Frame
 from compas_ifc.model import Model
 
 
-model = Model.template(storey_count=1)
+model = Model.template(storey_count=1, unit="m")
 
-# Default unit is mm
-box = Box.from_width_height_depth(5000, 5000, 50)
-sphere = Sphere(2000)
+box = Box.from_width_height_depth(5, 5, 0.05)
+sphere = Sphere(2)
 mesh = Mesh.from_obj(compas.get("tubemesh.obj"))
-mesh.scale(1000)
 
 storey = model.building_storeys[0]
 
 element1 = model.create("IfcWall", geometry=box, frame=Frame([0, 0, 0]), name="Wall", parent=storey)
-element2 = model.create("IfcRoof", geometry=mesh, frame=Frame([0, 0, 5000]), name="Roof", parent=storey)
-element3 = model.create(geometry=sphere, frame=Frame([0, 5000, 0]), name="Sphere", parent=storey, properties={"Custom_Pset": {"Custom_Property": "Custom Value"}})
+element2 = model.create("IfcRoof", geometry=mesh, frame=Frame([0, 0, 5]), name="Roof", parent=storey)
+element3 = model.create(geometry=sphere, frame=Frame([0, 5, 0]), name="Sphere", parent=storey, properties={"Custom_Pset": {"Custom_Property": "Custom Value"}})
 
 model.print_spatial_hierarchy(max_depth=5)
 model.show()