+++ src/compas_ifc/file.py
@@ -34,7 +34,6 @@ class IFCFile(object):
         else:
             self._file = ifcopenshell.open(filepath)
             print("IFC file loaded: {}".format(filepath))
-            self.model.update_linear_deflection()
 
         self._schema = ifcopenshell.ifcopenshell_wrapper.schema_by_name(self._file.schema)